Add tag filtering for black hole chests (#223)
